% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilica.r
\docType{package}
\name{basilica}
\alias{basilica}
\alias{basilica-package}
\title{Basilica}
\description{
Basilica allows you to turn your images and natural language text into features.
}
\section{Creating an API key}{


You can use basilica with our "SLOW_DEMO_KEY", which is an evaluation key with
a limit of 5,000 requests per week per IP address. You can create an API key
for free at www.basilica.ai, which will give you more requests.
}

\section{How many data points do I need?}{


For training your own models with embeddings provided by Basilica, you should
have around 1,000 data points. The more data points the better though. Some
models might have good results with less data, while others might need more.
}

\section{What do these features mean?}{


The features provided by Basilica are points in high-dimensional space where
two points that are considered similar. These embeddings are trained through
deep neural networks trained on a variety of tasks with millions of data
points. Go to https://www.basilica.ai/available-embeddings/ to read more about
our different embeddings.
}

\examples{
library(basilica)
connect("SLOW_DEMO_KEY")

sentences = list(
   "This is a sentence!",
   "This is a similar sentence!",
   "I don't think this sentence is very similar at all..."
)

embeddings <- embed_sentences(sentences)
print(dim(embeddings)) # 3 512
print(embeddings) # [[0.8556405305862427, ...], ...]

print(cor(embeddings[1,], embeddings[2,])) # 0.8048559
print(cor(embeddings[1,], embeddings[3,])) # 0.6877435

}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.basilica.ai}
  \item Report bugs at \url{https://github.com/basilica-ai/basilica-r-client/issues}
}

}
\author{
\strong{Maintainer}: Jorge Silva \email{jorge@basilica.ai}

}
