% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.r
\name{Connection_which_labeled}
\alias{Connection_which_labeled}
\title{Find the indices of particles labeled as identified}
\arguments{
\item{label}{one or more labels to search for}
}
\value{
zero or more indices of labeled items
}
\description{
Find the indices of particles labeled as identified
}
