% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{createConnection}
\alias{createConnection}
\title{createConnection}
\usage{
createConnection(auth_key = character(), server = character(),
  retries = numeric(), backoff_factor = numeric())
}
\arguments{
\item{auth_key}{Basilica API key}

\item{server}{Basilica server to point to (Default: `https://api.basilica.ai`)}

\item{retries}{Number of retries for any given request}

\item{backoff_factor}{How much to backoff}
}
\description{
Instantiates and returns a Basilica connection tied to a specific auth key and other connection parameters.
}
