% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilica.r
\name{embed_sentence}
\alias{embed_sentence}
\title{embed_sentence}
\usage{
embed_sentence(sentence = character(), model = "english",
  version = "default", timeout = 5)
}
\arguments{
\item{sentence}{Sentence or string}

\item{model}{Name of the image model you wish to use. (Default: `english`)}

\item{version}{Version of the image model you wish to use. (Default: `default`)}

\item{timeout}{Time (in seconds) before requests times out. (Default `5`)}
}
\value{
matrix
}
\description{
Get a vector of features for a sentence
}
