% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilica.r
\name{embed_sentences}
\alias{embed_sentences}
\title{embed_sentences}
\usage{
embed_sentences(sentences = list(), model = "english",
  version = "default", conn = environment(), timeout = 5)
}
\arguments{
\item{sentences}{List of sentences or strings}

\item{model}{Name of the image model you wish to use. (Default: `english`)}

\item{version}{Version of the image model you wish to use. (Default: `default`)}

\item{conn}{Basilica connection. Must be created with the `connect` function (Default: Global `basilica_connection`)}

\item{timeout}{Time (in seconds) before requests times out. (Default `5`)}
}
\value{
matrix
}
\description{
Get a vector of features for a list of sentences
}
