% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{class}
\name{Connection-class}
\alias{Connection-class}
\alias{Connection}
\title{Basilica Connection}
\description{
Constructor for a Basilica connection. Connections are tied to a specific auth_key and other connection parameters.
}
\section{Fields}{

\describe{
\item{\code{auth_key}}{Basilica API key}

\item{\code{server}}{Basilica server to point to (Default: `https://api.basilica.ai`)}

\item{\code{retries}}{Number of retries for any given request}

\item{\code{backoff_factor}}{How much to backoff}
}}

\section{Methods}{

\describe{
\item{\code{embed_image(image = character(), model = "generic", version = "default",
  timeout = 5)}}{Embed an image}

\item{\code{embed_images(images = character(), model = "generic", version = "default",
  timeout = 5)}}{Embed a list of images}

\item{\code{embed_sentence(sentence = character(), model = "english",
  version = "default", timeout = 5)}}{Embed a single sentence}

\item{\code{embed_sentences(sentences = list(), model = "english",
  version = "default", timeout = 5)}}{Embed a list of sentences}
}}

