% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basilica.r
\name{embed_image}
\alias{embed_image}
\title{embed_image}
\usage{
embed_image(image = raw(), model = "generic", version = "default",
  timeout = 5)
}
\arguments{
\item{image}{Path to an image (JPEG or PNG)}

\item{model}{Name of the image model you wish to use. (Default: `generic`)}

\item{version}{Version of the image model you wish to use. (Default: `default`)}

\item{timeout}{Time (in seconds) before requests times out. (Default `5`)}
}
\value{
matrix
}
\description{
Get a vector of features for an image
}
